requirements.txt
fastapi
python-dotenv
uvicorn
openai
elevenlabs
